+++ scripts/test_interferometric_analysis.py
@@ -7,12 +7,11 @@ from perseo_quality.interferometric_analysis.graphical_output import generate_co
 from perseo_quality.interferometric_analysis.support import coherence_histograms_to_netcdf
 
 from sct.analyses import interferometry as interf
-from sct.configuration.logger import ConsoleHandler, enable_quality_logger, sct_logger
+from sct.configuration.logger import enable_quality_logger, sct_logger
 from sct.configuration.config import SCTInterferometricAnalysisConfig
 
 # setup custom logger
 enable_quality_logger()
-sct_logger.addHandler(ConsoleHandler())
 sct_logger.setLevel("INFO")